This commit is an attempt to continue moving gently towards the JSR77 state mechanism without committing to any MBean backbone choise etc.
This commit has some of the ideas from the patch submitted by gianny_damour@hotmail.com
Specifically the addition of the JSR77 state change notifications to the State class.
However I have left his full notification mechanism out for now, as I am unclear
if this is the way we are going.  I have put and abstract notification method
onto AbstractComponent, so it will be simple to add the real notification mechanism when
we know what that is.

The rest of the commit is more in the same theme. Gently moving towards JSR77.
I have extracted the state stuff from Component into a StateManageable interface
and an AbstractStateManageable.

Unfortunately JSR77 wants it's state as an int, so getState now returns
an int.  But getStateInstance is there to get the object itself.  I have
put a bit more toInt an fromInt support into the State class.

For symmetry I created the AbstractContainer class, but it does nothing
at this stage.  It needs the startRecursive handling moved to it once
we have a real Container interface (one which you can iterate over
the Components would be good).



Submitted by: Greg Wilkins
Reviewed by: Nameless the cat


git-svn-id: https://svn.apache.org/repos/asf/geronimo/trunk@43779 13f79535-47bb-0310-9956-ffa450edef68