Added GBean attribute contextPriorityClassLoader.   If set true then the servlet spec recommendation is used for classloading.
If set to false ( the default ) then it is not.   We should be able to
directly use the geronimo loader as the context loader, but currently
it does not appear to find the WEB-INF/lib and WEB-INF/classes stuff, so
we are still making a jetty loader.

Also added the attribute to the DConfigBean, but currently it does not
appear that this is actualy being used by anything other than the tests?


git-svn-id: https://svn.apache.org/repos/asf/geronimo/trunk@45007 13f79535-47bb-0310-9956-ffa450edef68