Patch from Srinath....Here are the notes from his message in GERONIMO-269:
I have fix the test cases.. I am working around the compiling problem.
Right now the user should use ews to generate the module and Only then
it can be deployed. I am generating and compile the module with ews inside
the maven.xml file.

The patch is attached .. eclipse does not a samples dir to the patch as it
is not cvs managed I zip it and add it. plase drop it in to the
modules/axis/src

The code of the Jetty and the OpenEJB modules changing quite regulary (it
was twice for last two weeks) and that is how the AxisGbean test fails. I
add notes at the code specifying the palces where we should check if the
test fails.


git-svn-id: https://svn.apache.org/repos/asf/geronimo/trunk@45737 13f79535-47bb-0310-9956-ffa450edef68