Patch from Srinath: =================================================== with the new patch fix the class loading stuff. Now this complete geronimo supports for the Simple type web serivices. The WebServiceTest now has code to do the actual web service invocation tests.
There is a class loading problem for the Complex type web services in the
openejb service invocation. I add a test case to cover this but exclude
it.

I need to check it in the new openEJB cvs and get help from the open ejb
ppl if there is a problem.
Thanks
--Srinath
=======================================================


git-svn-id: https://svn.apache.org/repos/asf/geronimo/trunk@45780 13f79535-47bb-0310-9956-ffa450edef68