Latest fixes from Srinath and some code cleanup from me.
===========================================
Most of the cooments from the community is incoperated with this
patch. This patch remove the referances to the kernel and remove
hardcoded obj names. add a POJO sample and it's test case is working.
there is a problem with the EJB test case will fix that and send a
patch.
===========================================



git-svn-id: https://svn.apache.org/repos/asf/geronimo/trunk@56419 13f79535-47bb-0310-9956-ffa450edef68