Kernel is not totally decoupled from JMX.  By default the kernel will no longer boot an MBeanServer.  The only place we boot the MBean server is when we are booting a full blown long running Geronimo server.
Introduced GBeanRegistry which manages the map ObjectName to GBeanInstance.

Added BasicGBeanRegistry which manages the map with no dependency on JMX.

Added JMXGBeanRegistry which creates an MBeanServer on startup.  It mounts a o.a.g.kernel.jmx.GBeanMBean into an MBeanServer for each GBeanInstance registered.  The a o.a.g.kernel.jmx.GBeanMBean does not interact directly with GBeanInstances, instead it simply redirects all calls to a GBean.  GBeanMBean in the o.a.g.gbean.jmx package is just a proxy for a GBeanData, and should be removed soon.

Added KernelDelegate which replaces the MBeanServerFactory use on the client side to create a proxy to the server side geronimo kernel.

Added MBeanServerDelegate which is used to provide apis such as the JSR 88 MEJB which needs a something that implements MBeanServer.  The MBeanServerDelegate only implements the operations of the MBeanServer that easily map to Kernel.  All unimplemented methods throw a SecurityException.

Moved Lifecycle* classes to o.a.g.kernel.lifecycle


git-svn-id: https://svn.apache.org/repos/asf/geronimo/trunk@109957 13f79535-47bb-0310-9956-ffa450edef68