Change GBean registries to key off GBeanName rather than ObjectName. To reduce impact, the kernel API is currently unchanged and the kernel maps names accordingly; this will be removed when the API is updated.
The JMXGBeanRegistry now uses an injected MBeanServer rather than
creating one itself; client code that creates this type of registry
has been updated to create the MBeanServer as well.

The GBeanDataRegistry (which holds GBeanData not GBeanInstance instances)
is no longer a GBeanRegistry but its own class. This is only used by
DeploymentContext.



git-svn-id: https://svn.apache.org/repos/asf/geronimo/trunk@154941 13f79535-47bb-0310-9956-ffa450edef68