Allow remote class loading via RMI annotation for deployment. This includes:
o the definition of a ClassLoaderServer used to export and unexport 
ClassLoaderInfo. This later is a simple wrapper around ClassLoader.

Provide an implementation of ClassLoaderServer using HTTP as the communication
protocol.

o define ConfigurationClassLoader, a specific ClassLoader for Configuration,
which adds two new attributes on top of the URLClassLoader type: ID, which is 
the ID of the defining Configuration; and classLoaderServerURLs, which is
an array of URLs defining the locations from which the Configuration classes
are services by a ClassLoaderServer.

o impact RMIClassLoaderSpiImpl such that it uses the classLoaderServerURLs
defined by a ConfigurationClassLoader to provide the codebase of serialized
classes.

o add the org/apache/geronimo/RemoteClassLoadingDeployer configuration, which
exports the ConfigurationClassLoader of the org/apache/geronimo/Server and
org/apache/geronimo/RuntimeDeployer configurations. 

o add a policy file, which can be used along with deployer.jar to enable
remote class loading.

This policy file is "opened" and needs to be updated to reflect the
execution environment.


git-svn-id: https://svn.apache.org/repos/asf/geronimo/trunk@158589 13f79535-47bb-0310-9956-ffa450edef68