A couple of changes to get document style working, at least for POJO WS.
o update JettyPOJOWebServiceHolder such that it registers during start-up the
TypeDesc of the exposed service. This also includes a new dependency to
geronimo-axis from geronimo-jetty.

o add a TypeDescInfo class wrapping all the information required to recreate
a TypeDesc. As a matter of fact TypeDesc cannot be serialized "as is" due
to the contained property descriptors, which are not serializable. An Axis
patch is required to fix this problem.

o update HeavyweightOperationDescBuilder such that parameters and return
value of document style operations are "correctly" mapped. In the case of
document style operations, the names of the parameters and return values are 
the names of the elements defined by the message part.

o fix AxisServiceBuilder.buildOperations such that the soap:binding style is 
provided as the default style for the nested soap:operation.

o Definition.getTypes() may return null. This is not specified by the API;
yet, I have observed this behavior.


git-svn-id: https://svn.apache.org/repos/asf/geronimo/trunk@161304 13f79535-47bb-0310-9956-ffa450edef68