A couple of changes to get lightweight types mapping working and properly register the TypeDesc from the client invoking the WS.
This includes:

o implement TypeDesc builder for lightweight mappings;

o make ReadOnlyServiceDesc Externalizable. The custom deserialization is
used to register the underlying TypeDesc prior to deserialize the contained 
JavaServiceDesc. This is required as JavaServiceDesc contains a bunch of 
BeanXXXFactory requiring the relevant TypeDesc to be registered in order to 
properly recreate their states; and

o remove the dependency to geronimo-axis from geronimo-jetty and roll-back
my previous fix to JettyPOJOWebServiceHolder as it is now redundant. My
previous fix was registering the TypeDesc prior to deserialize the 
JavaServiceDesc contained by WebServiceContainer.


git-svn-id: https://svn.apache.org/repos/asf/geronimo/trunk@161563 13f79535-47bb-0310-9956-ffa450edef68