Slightly change the way the FieldDesc of a TypeDesc are built:
o Lightweight type info modification: use the raw QName of the element 
specified in the SchemaType associated to the TypeDesc being built as the 
FieldDesc XmlName.
The previous implementation was using the namespace of the SchemaType and
the local part name of the element.

o Heavyweight type info modification: try "" for the namespace and the
local part name of the element as specified in the explicit mapping. If 
the resulting QName exists in the SchemaType associated to the type info
being built, then use this name.
Otherwise, try the namespace of the SchemaType associated to the type info
being built.

o remove the FieldDesc.setJavaType calls as Axis does not use then.

Also, fix a couple of copyrights.


git-svn-id: https://svn.apache.org/repos/asf/geronimo/trunk@164247 13f79535-47bb-0310-9956-ffa450edef68