o update LightweightTypeInfoBuilder to mirror what HeavyweightTypeInfoBuilder is doing for description of arrays.
o if the type of a ParameterDesc is an element mapped to a simple type,
then the element needs to be registered by the type system.

o clean-up the way serializer/deserializer are retrieved.

o rely on axis to retrieve the serializer/deserializer pairs for 
a given Java type and XML type.


git-svn-id: https://svn.apache.org/repos/asf/geronimo/trunk@178770 13f79535-47bb-0310-9956-ffa450edef68