Intermediate check-in for in-place deployment.
At this stage, in-place deployment seems to work as expected on the server-side 
(I have been able to run jUDDI as an in-place configuration); however, it is not
yet possible to pass in an in-place deployment flag to the deployer command line.

Fix also a minor bug in JettyModuleBuilder which was not dropping the leading
'/' from the contextRoot, if any, when building the artefact name.

In a few words, when a in-place configuration is built, an
"inPlaceConfiguration.txt" file is created beside the standard "config.ser".
This file contains the full path of the in-place configuration. I have decided
to make that a simple file beside config.ser such that a user can easily create
this file and have G automatically mount the specified location. Also, it avoids
to load the config.ser to get the in-place configuration.

To give it a quick test, move the content of 
geronimo/uddi-jetty/1.1-SNAPSHOT/uddi-jetty-1.1-SNAPSHOT, except META-INF dir,
to any directory. Create META-INF/inPlaceConfiguration.txt pointing to this location.


git-svn-id: https://svn.apache.org/repos/asf/geronimo/branches/1.1@393489 13f79535-47bb-0310-9956-ffa450edef68