Fix build following recent Jetty6 AbstractSessionManager updates.
There are two core changes:
1. update ClusteredSessionHandler to leverage the new API provided
by AbstractSessionManager, which simplifies the implementation and
remove a hack around the initialization of the session attributes
following a session migration.
2. move pre-handle logic from ServletHandler to SessionHandler.

Also, fix a lame type in WADIClusteredPreHandlerFactory: the
expose interface is PreHandlerFactory and not SessionHandlerFactory.

This fixes GERONIMO-2677 - HttpSession Relocation - Sticky load-balancing via HTTP Cookie


git-svn-id: https://svn.apache.org/repos/asf/geronimo/server/trunk@493201 13f79535-47bb-0310-9956-ffa450edef68