Second attempt to change the way the online/offline deployers and the JSR88 deployment driver work. The first attempt was breaking the TCK amd the eclipse plugin.
Online deployer, i.e. deployer.jar, boots a Kernel to load its dependencies,
 e.g. geronimo-deploy-tool, and registers the available ModuleConfigurers with 
the DeploymentManager.

ModuleConfigurers to be registered are loaded by the persistent configuration 
list jsr88-configurer-config.xml.

In the case of an offline deployment, the online deployer starts the 
offline-deployer configuration within the same Kernel. In turn, the 
offline-deployer configuration starts a list of configurations to register 
the available module builders.

Add a log4j configuration for the online deployer.

DeploymentFactoryBootstrapper is the new JSR88 deployment driver. It boots a 
kernel; starts the configuration list jsr88-configurer-config.xml; retrieves
the "actual" DeploymentFactory implementation from the kernel; and delegates
to this retrieved imoplementation.
The JSR88 JAR driver is now named jsr88-deploymentfactory.jar.


This fixes:
* GERONIMO-2794 - Improve online deployer to register ModuleConfigurers from the repository; and
* GERONIMO-2767 - Minimize side effects of the offline deployer


git-svn-id: https://svn.apache.org/repos/asf/geronimo/server/trunk@508724 13f79535-47bb-0310-9956-ffa450edef68