GERONIMO-2805   tribes clustering support for tomcat
copied the tomcat clustering support from server/branches/1.2 and refactored it to work with catalina-ha
and tribes from tomcat6.  Added InterceptorGBean, ChannelGBean, and morphed MessageListenerGBean into
ClusterListenerGBean to correspond with the new arch in tomcat 6. I will attach a geronimo config.xml to
the JIRA that can be used to startup a clustered server.  Reenabled the clustering support in EngineGBean
and the various o.a.g.tomcat classes.



git-svn-id: https://svn.apache.org/repos/asf/geronimo/server/trunk@511136 13f79535-47bb-0310-9956-ffa450edef68