GERONIMO-2955 enable the Jasper servlet to find TLDs in Geronimo's MultiParentClassLoader
When a JSP contains a URI reference to a tag library Jasper scans the jar files
in the classloader hierarchy looking for a TLD file that defines the tags for that
URI.  The method it uses to navigate the classloader hierarchy is not compatible
with Geronimo's MultiParentClassLoader.

Jasper allows you to override its default JSP options although it's not very
straightforward to do so.  I could not directly extend the default impl
of the Options interface because it is declared final.  Also, most of the
important bits of TldLocationsCache are private so I made a copy of it and
adjusted the appropriate methods.



git-svn-id: https://svn.apache.org/repos/asf/geronimo/server/trunk@517712 13f79535-47bb-0310-9956-ffa450edef68