Should have been ci as part of r519908.
* Do not start jsr88-cli within the server. It defines a GBean implementing
the PluginInstaller type, GBean named ModuleConfigurerRegistry, which
is not intended to be used within a server as the other GBeans of this 
module. This causes a NPE during the installation of plugins via the command
line deployer: CommandInstallCAR tries to execute startInstall
against all the GBeans implementing the PluginInstaller type and this
invokation fails with a NPE against ModuleConfigurerRegistry.

* Push up the DeploymentFactoryImpl implementation to BaseDeploymentFactory,
such that DeploymentFactoryWithKernel can extend it instead of 
DeploymentFactoryImpl. This refactoring is done to ensure that 
DeploymentFactoryWithKernel does not register a DeploymentFactoryImpl
instance with DeploymentFactoryManager (done via a static block).

This fixes GERONIMO-2983 Installation of plugins via the command line is 
broken.

git-svn-id: https://svn.apache.org/repos/asf/geronimo/server/trunk@519919 13f79535-47bb-0310-9956-ffa450edef68