Drop ReplicationManager and ReplicaStorage GBeans. WADI's API has slightly changed and it is no more possible to plug-in such strategies via this approach.
Pass in the ClassLoader of the configuration which is defining the
BasicWADISessionManager GBean to the WADI StackContext. This ClassLoader
is used to unserialize replicated sessions. Previously, replicated sessions
were not unserialized by their storages. They are no fully re-instantiated so 
that field or method level changes can be applied to them.


git-svn-id: https://svn.apache.org/repos/asf/geronimo/server/trunk@558233 13f79535-47bb-0310-9956-ffa450edef68