Add a convenience PropertyEditor, JavaBeanXmlAttributeEditor, which is intended to be used as a base class for property editors of JavaBeans defined as gbean xml-attributes. This base property editor is able to skip the "xml-serialization" of properties whose read methods are annotated with @DoNotPersist. It is also able to encrypt prior xml-serialization the properties whose read methods are annotated with @EncryptOnPersist.
It is now possible to declare a propertyEditor on a config.xml gbean attribute.
This property editor is used to retrieve the value of the attribute based
on its mixed-content. Combined with the above convenience PropertyEditor, it
is now easy to override gbean attributes defined by Java xml-attribute in
configuration plans.

Write a BasicExtendedJMXConnectorInfoEditor to allow the override of the 
extendedJMXConnectorInfo attribute of BasicNodeInfo. Users can now easily
declare new nodes within the clustering configuration.

Fixes GERONIMO-3610 - Allows the override of XML JavaBean attribute in 
config.xml


git-svn-id: https://svn.apache.org/repos/asf/geronimo/server/trunk@596542 13f79535-47bb-0310-9956-ffa450edef68