When distributing w/o the explicit selection of a target, we only distribute to the first target returned by DeploymentManager. This first target is the default target, or configuration store, and users can specify an abstract name query (in EditableKernelConfigurationManager) to identify this default target.
Now that multiple configuration stores per server are better handled 
out-of-the-box, we can re-enable the "clustered" configuration stores.

Also, let's more the rmock dependency to testsupport-common so that we
can use it from all our tests.

This fixes GERONIMO-3612 - When no target configuration store is explicitly 
specified while installing a configuration, the configuration should be 
installed to a default configuration store


git-svn-id: https://svn.apache.org/repos/asf/geronimo/server/trunk@596987 13f79535-47bb-0310-9956-ffa450edef68