1. Add SessionManagerListener to the clustering API. This allows clients of local SessionManagers to receive callbacks when corresponding remote SessionManagers join or leave the cluster.
2. Arbitrary WADI clustered services can now be registered: distributed
services are wrapped within ClusteredServiceHolder GBeans which are injected
to BasicWADISessionManager and registered with the underlying WADI 
ServiceRegistry.

3. Add support for clustering of SFSBs.

The structure mirrors the one applied for the clustering of Jetty and
Tomcat, i.e. two modules and two configurations are added:

- geronimo-openejb-clustering-wadi: defines a ClusteredRPCContainer
implementation for SFSBs and infrastructure components maintaining a 
bookkeeping of the connection addresses a clustered SFSB client can connect
to. The monitoring components, i.e. NetworkConnectorMonitor and 
NetworkConnectTracker, works as follows:
o When an OpenEJB NetworkConnector service starts, NetworkConnectMonitor
registers the address of this NC with the clustered service 
NetworkConnectorTracker and this for all the locally running clustered SFBS 
deployments.
o When an OpenEJB NC services stops, NCM executes an unregisteration reverting
the effects of a NC service start.
o When a clustered SFSB deployments starts, NCM registers its  deployment id
with the clustered service NCT for all the locally running NCs addresses.
o When a clustered SFSB deployment stops, NCM executes an unregisters reverting
the effects of a deployement start.
o When a node hosting an EJB module having clustered SFSBs stops or fails, NCM
updates the local NCT so that all the connection addresses registered for
all the clustered SFSB deployments of this node are removed.

- geronimo-openejb-clustering-builder-wadi: defines a builder to handle the
new substitution group element 'clustering' added to geronimo-openejb-2.0.xsd.
The clustering builder works nearly the same than for Jetty and Tomcat except
for the following specificities:
o A (WADI) clustered NCT service is registered.
o The container id for the clustered SFSBs is set to the Geronimo provided
clustered SFSB container implementation.

- openejb-clustering-wadi: declares Geronimo provided clustered SFSB
container implementation.

- openejb-clustering-builder-wadi: declares clustering builder.


git-svn-id: https://svn.apache.org/repos/asf/geronimo/server/trunk@630370 13f79535-47bb-0310-9956-ffa450edef68